- kvector: float -> double - implemented zoom tool
Note: The rulers and scrollbars do not understand the zooming yet. I'm going to fix that later.

svn path=/trunk/koffice/; revision=27374

+++ kimageshop/kvector.cc
@@ -22,7 +22,7 @@
 
 KVector& KVector::normalize()
 {
-  float	length, ilength;
+  double length, ilength;
 
   length = m_x*m_x + m_y*m_y + m_z*m_z;
   length = sqrt (length);
@@ -39,7 +39,7 @@ KVector& KVector::normalize()
 
 KVector& KVector::crossProduct(const KVector &v)
 {
-  float x,y,z;
+  double x,y,z;
 
   x = m_y*v.m_z - m_z*v.m_y;
   y = m_z*v.m_x - m_x*v.m_z;
